+++ src/commands/tag.ts
@@ -33,7 +33,7 @@ export const sync_tag = async (commandEvent: CommandEvent) => {
     commandEvent.command
   );
   try {
-    console.log("sync_tag");
+    // console.log("sync_tag");
 
     const new_bench_time = new Date().toISOString();
     const bench_time_key = "bench_time_tag";
@@ -100,7 +100,7 @@ export const sync_tag = async (commandEvent: CommandEvent) => {
           );
           result.created++;
         } catch (e: any) {
-          console.log("Create Tag Failed >> ", e?.response, body);
+          // console.log("Create Tag Failed >> ", e?.response, body);
           failed_docs_report.push({
             method: "create",
             doc: body,
@@ -122,7 +122,7 @@ export const sync_tag = async (commandEvent: CommandEvent) => {
           );
           result.updated++;
         } catch (e: any) {
-          console.log("Update Tag Failed >> ", e?.response?.data, body);
+          // console.log("Update Tag Failed >> ", e?.response?.data, body);
           failed_docs_report.push({
             method: "update",
             doc_id: repzo_tag?._id,