+++ src/test/actions/create_invoice.ts
@@ -182,6 +182,7 @@ Actions(
     isBase64Encoded: false,
   },
   {
+    app_id: "",
     repzoEndPoint: "",
     serviceEndPoint: "",
     env: "staging",