+++ src/test/actions/create_return_invoice.ts
@@ -194,6 +194,7 @@ Actions(
     isBase64Encoded: false,
   },
   {
+    app_id: "",
     repzoEndPoint: "",
     serviceEndPoint: "",
     env: "staging",