+++ changelog.md
@@ -8,6 +8,8 @@
 - New Commands: [warehouse, rep, tag, tax, measureunit, measureunit-family, category, channel, paymentTerm, bank, priceList, priceListItem, product, disabled_product, client, disabled_client, adjust_inventory] @maramalshen
 - New Actions: [invoice, return_invoice, proforma, payment, transfer] @maramalshen
 - fix bug in sync_client @maramalshen
+- add detail with the action.sync_id
+- if create transfer was failed then send command: Adjust Inventory (optionals)
 
 ### Changed